+++ src/lib/Components/Bidding/Screens/ConfirmBid.tsx
@@ -185,6 +185,7 @@ export class ConfirmBid extends React.Component<ConfirmBidProps, ConfirmBidState
         addressCity: billingAddress.city,
         addressState: billingAddress.state,
         addressZip: billingAddress.postalCode,
+        addressCountry: billingAddress.country.shortName,
       })
 
       commitMutation(this.props.relay.environment, {
@@ -367,7 +368,8 @@ export class ConfirmBid extends React.Component<ConfirmBidProps, ConfirmBidState
               <SerifSemibold14>Lot {lot_label}</SerifSemibold14>
 
               <SerifItalic14 color="black60" textAlign="center">
-                {artwork.title}, <Serif14>{artwork.date}</Serif14>
+                {artwork.title}
+                {artwork.date && <Serif14>, {artwork.date}</Serif14>}
               </SerifItalic14>
             </Flex>